+++ ShapeLib/Parameters/ShapeParameter.cs
@@ -2,10 +2,10 @@ namespace ShapeLib
 {
     public class ShapeParameter
     {
-        public float DefaultValue { get; }
+        public double DefaultValue { get; }
         public IParameterValidation Validation { get; }
 
-        public ShapeParameter(float defaultValue, IParameterValidation validation)
+        public ShapeParameter(double defaultValue, IParameterValidation validation)
         {
             DefaultValue = defaultValue;
             Validation = validation;
