+++ ShapeLib/Parameters/Validation/AlwaysValid.cs
@@ -3,7 +3,7 @@ namespace ShapeLib
     public class AlwaysValid : IParameterValidation
     {
         public static readonly AlwaysValid Instance = new AlwaysValid();
-        public bool IsValid(float value) => true;
-        public void ThrowIfNotValid(float value) { }
+        public bool IsValid(double value) => true;
+        public void ThrowIfNotValid(double value) { }
     }
 }
