+++ ShapeLib/Parameters/Validation/IParameterValidation.cs
@@ -2,7 +2,7 @@ namespace ShapeLib
 {
     public interface IParameterValidation
     {
-        public bool IsValid(float value);
-        void ThrowIfNotValid(float value);
+        public bool IsValid(double value);
+        void ThrowIfNotValid(double value);
     }
 }
