+++ ShapeLib/Shapes/IShape.cs
@@ -5,6 +5,6 @@ namespace ShapeLib
         public string ShapeName { get; }
         public IReadOnlyDictionary<string, ShapeParameter> ShapeParametersScheme { get; }
        // public void SetParameters(Dictionary<string, float> shapeParameters);
-        void SetParameters(Dictionary<string, float> parameters);
+        void SetParameters(Dictionary<string, double> parameters);
     }
 }
