+++ ShapeLib/Shapes/ShapeQuad.cs
@@ -11,19 +11,19 @@ namespace ShapeLib
             { SideB, new ShapeParameter(0, ValidGreaterThan.Zero)} 
         };  
         
-        private float _sideA;
-        private float _sideB;
-        private float _area;
+        private double _sideA;
+        private double _sideB;
+        private double _area;
         
         public IReadOnlyDictionary<string, ShapeParameter> ShapeParametersScheme => Scheme;
         public string ShapeName => GetType().Name;
         
-        public void SetParameters(Dictionary<string, float> shapeParameters)
+        public void SetParameters(Dictionary<string, double> shapeParameters)
         {
             _sideA = shapeParameters[SideA];
             _sideB = shapeParameters[SideB];
             _area = _sideA * _sideB;
         }
-        public float GetArea() => _area;
+        public double GetArea() => _area;
     }
 }
